+++ include/build.h
@@ -2,5 +2,6 @@
  * Author: sascha_lammers@gmx.de
  */
 #pragma once
-#define __BUILD_NUMBER "10135"
-#define __BUILD_NUMBER_INT 10135
+#define __BUILD_NUMBER "10258"
+#define __BUILD_NUMBER_INT 10258
+